use full precision whisper, add retrieval

+++ prepare/preprocess_ppg.py
@@ -21,7 +21,7 @@ def load_model(path) -> Whisper:
     del model.encoder.blocks[cut:]
     model.load_state_dict(checkpoint["model_state_dict"], strict=False)
     model.eval()
-    model.half()
+    #model.half()
     model.to(device)
     return model
 
@@ -31,7 +31,8 @@ def pred_ppg(whisper: Whisper, wavPath, ppgPath):
     audln = audio.shape[0]
     ppgln = audln // 320
     audio = pad_or_trim(audio)
-    mel = log_mel_spectrogram(audio).half().to(whisper.device)
+    #mel = log_mel_spectrogram(audio).half().to(whisper.device)
+    mel = log_mel_spectrogram(audio).to(whisper.device)
     with torch.no_grad():
         ppg = whisper.encoder(mel.unsqueeze(0)).squeeze().data.cpu().float().numpy()
         ppg = ppg[:ppgln,]  # [length, dim=1280]
